src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-reprojection-recalc-expected.txt

CONSOLE ERROR: line 16: Uncaught TypeError: host.createShadowRoot is not a function
Redistribution into same position should not cause style recalc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.internals is defined.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
